fix(shepherd): identifying what cron task is blowing up, breaking up the modules a bit
